move core.testutil -> dev.testutil
