+++ pywincffi/dev/testutil.py
